Update loot_trading_cards.json
Felt bad abandoning this halfway through so I finished some more of it. none of it is tested in any way so I wouldn't be surprised if something's not right. Might finish the rest, might not, who knows. Removed the "profession" cards and shifted all the numbers down because I thought they were boring.